Dashboard/requirements.txt
flask  
flask-sqlalchemy  
flask-socketio  
paho-mqtt  
pytest  
pytz  
requests  
sqlalchemy

